requirements.txt
projz.py
aiohttp
aiofiles